backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
matplotlib==3.10.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# File handling and storage
python-magic-bin==0.4.14
supabase==2.7.4

# PDF forensics
# pdfid==1.1.3  # This is just a Python wrapper, we need the binary
pikepdf==8.6.0
PyPDF2==3.0.1

# Image forensics
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Machine Learning
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2

# Error Level Analysis
imageio==2.33.0
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
# pydantic==2.5.0
pydantic<2
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
cryptography>=41.0.0,<46.0.0

# Logging and monitoring
structlog==24.1.0

# Market Intelligence & Fact-Checker Module
nsepython==2.97
transformers==4.35.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
lxml==4.9.3